// src/generated/graphql.ts
export type Maybe<T> = T | null
// ...
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A date string, such as 2007-12-03, compliant with the `full-date` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: any
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: any
  EmailAddress: any
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any
  NonNegativeFloat: any
  /** Integers that will have a value of 0 or more. */
  NonNegativeInt: any
  PhoneNumber: any
  /** Integers that will have a value greater than 0. */
  PositiveInt: any
}
// ...
export type CreateSubscriptionPlanInput = {
  identifier: Scalars['String']
  meteredVariables?: Maybe<Array<SubscriptionPlanMeteredVariableInput>>
  name?: Maybe<Scalars['String']>
  periods: Array<SubscriptionPlanPeriodInput>
  instanceId: Scalars['ID']
}
// ...
export type Grid = {
  __typename?: 'Grid'
  createdAt: Scalars['DateTime']
  hasVersion: Scalars['Boolean']
  id: Scalars['ID']
  language: Scalars['String']
  meta?: Maybe<Array<KeyValuePair>>
  metaProperty?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
  rows: Array<GridRow>
  instanceId: Scalars['ID']
  updatedAt?: Maybe<Scalars['DateTime']>
  version?: Maybe<VersionedRecordInfo>
}
// ...
export type Image = {
  __typename?: 'Image'
  altText?: Maybe<Scalars['String']>
  caption?: Maybe<RichTextContent>
  key: Scalars['String']
  meta?: Maybe<Array<KeyValuePair>>
  metaProperty?: Maybe<Scalars['String']>
  mimeType?: Maybe<Scalars['String']>
  url?: Maybe<Scalars['String']>
  variants?: Maybe<Array<ImageVariant>>
}
// ...
export type KeyValuePair = {
  __typename?: 'KeyValuePair'
  key: Scalars['String']
  value?: Maybe<Scalars['String']>
}
// ...
export type Language = {
  __typename?: 'Language'
  code: Scalars['String']
  name: Scalars['String']
  system: Scalars['Boolean']
}
// ...
export type Shape = {
  __typename?: 'Shape'
  components?: Maybe<Array<ShapeComponent>>
  createdAt: Scalars['DateTime']
  /** @deprecated Shape IDs have been deprecated in favor of human readable identifiers. Querying for them will be removed in a future release. */
  id?: Maybe<Scalars['ID']>
  identifier: Scalars['String']
  itemCount: Scalars['Int']
  items?: Maybe<Array<Item>>
  meta?: Maybe<Array<KeyValuePair>>
  metaProperty?: Maybe<Scalars['String']>
  name: Scalars['String']
  instanceId: Scalars['ID']
  type: ShapeType
  updatedAt?: Maybe<Scalars['DateTime']>
}
// ...
export type SubscriptionPlan = {
  __typename?: 'SubscriptionPlan'
  createdAt: Scalars['DateTime']
  identifier: Scalars['String']
  meteredVariables?: Maybe<Array<SubscriptionPlanMeteredVariable>>
  name?: Maybe<Scalars['String']>
  periods?: Maybe<Array<SubscriptionPlanPeriod>>
  instance: Instance
  instanceId: Scalars['ID']
  updatedAt?: Maybe<Scalars['DateTime']>
}

export type SubscriptionPlanMeteredVariable = {
  __typename?: 'SubscriptionPlanMeteredVariable'
  id: Scalars['ID']
  identifier: Scalars['String']
  name?: Maybe<Scalars['String']>
  unit: Scalars['String']
}

export type SubscriptionPlanMeteredVariableInput = {
  id?: Maybe<Scalars['ID']>
  identifier: Scalars['String']
  name?: Maybe<Scalars['String']>
  unit: Scalars['String']
}
// ...
export type SubscriptionPlanPeriod = {
  __typename?: 'SubscriptionPlanPeriod'
  id: Scalars['ID']
  initial?: Maybe<SubscriptionPlanPhase>
  name: Scalars['String']
  recurring: SubscriptionPlanPhase
}

export type SubscriptionPlanPeriodInput = {
  id?: Maybe<Scalars['ID']>
  initial?: Maybe<SubscriptionPlanPhaseInput>
  name: Scalars['String']
  recurring: SubscriptionPlanPhaseInput
}

export type SubscriptionPlanPeriodReferenceInput = {
  id: Scalars['ID']
  initial?: Maybe<SubscriptionPlanPriceInput>
  recurring: SubscriptionPlanPriceInput
}

export type SubscriptionPlanPhase = {
  __typename?: 'SubscriptionPlanPhase'
  period: Scalars['Int']
  unit: SubscriptionPeriodUnit
}

export type SubscriptionPlanPhaseInput = {
  period: Scalars['Int']
  unit: SubscriptionPeriodUnit
}
// ...
export type Instance = {
  __typename?: 'Instance'
  authenticationMethod?: Maybe<InstanceAuthenticationMethod>
  availableLanguages?: Maybe<Array<Language>>
  createdAt: Scalars['DateTime']
  defaults: InstanceDefaults
  grids?: Maybe<Array<Grid>>
  id: Scalars['ID']
  identifier: Scalars['String']
  isActive: Scalars['Boolean']
  isTrial: Scalars['Boolean']
  logo?: Maybe<Image>
  meta?: Maybe<Array<KeyValuePair>>
  metaProperty?: Maybe<Scalars['String']>
  metrics?: Maybe<InstanceMetrics>
  name: Scalars['String']
  rootItemId: Scalars['ID']
  shapes?: Maybe<Array<Shape>>
  staticAuthToken?: Maybe<Scalars['String']>
  topics?: Maybe<Array<Topic>>
  tree?: Maybe<Array<TreeNode>>
  updatedAt?: Maybe<Scalars['DateTime']>
  users?: Maybe<Array<UserInstanceRole>>
  vatTypes?: Maybe<Array<VatType>>
  webhooks?: Maybe<Array<Webhook>>
}
// ...
export type InstanceAuthenticationMethod = {
  __typename?: 'InstanceAuthenticationMethod'
  catalog?: Maybe<AuthenticationMethod>
  search?: Maybe<AuthenticationMethod>
}

export type InstanceAuthenticationMethodInput = {
  catalog?: Maybe<AuthenticationMethod>
  search?: Maybe<AuthenticationMethod>
}

export type InstanceDefaults = {
  __typename?: 'InstanceDefaults'
  currency: Scalars['String']
  language: Scalars['String']
}

export type InstanceDefaultsInput = {
  currency?: Maybe<Scalars['String']>
  language?: Maybe<Scalars['String']>
}

export type InstanceMetrics = {
  __typename?: 'InstanceMetrics'
  apiCalls: ApiCallMetrics
  bandwidth: BandwidthUsageMetrics
  items: ItemMetrics
  orders: OrderMetrics
  shapes: ShapeMetrics
  users: UserMetrics
  webhooks: WebhookMetrics
}
// ...
export type Topic = {
  __typename?: 'Topic'
  ancestors?: Maybe<Array<Topic>>
  children?: Maybe<Array<Topic>>
  createdAt: Scalars['DateTime']
  descendants?: Maybe<Array<Topic>>
  id: Scalars['ID']
  items?: Maybe<Array<Item>>
  language?: Maybe<Scalars['String']>
  name?: Maybe<Scalars['String']>
  parent?: Maybe<Topic>
  parentId?: Maybe<Scalars['ID']>
  path: Scalars['String']
  instance: Instance
  instanceId: Scalars['ID']
  updatedAt?: Maybe<Scalars['DateTime']>
}
// ...
export type TreeNode = {
  __typename?: 'TreeNode'
  ancestors?: Maybe<Array<TreeNode>>
  childCount?: Maybe<Scalars['Int']>
  children?: Maybe<Array<TreeNode>>
  identifiers?: Maybe<Array<TreeNodeIdentifier>>
  item?: Maybe<Item>
  itemId: Scalars['ID']
  parent?: Maybe<TreeNode>
  parentId?: Maybe<Scalars['ID']>
  path?: Maybe<Scalars['String']>
  position?: Maybe<Scalars['PositiveInt']>
  siblings?: Maybe<Array<TreeNode>>
  versionLabel: VersionLabel
}
// ...
export type UserInstanceRole = {
  __typename?: 'UserInstanceRole'
  role?: Maybe<UserRole>
  instance: Instance
  instanceId: Scalars['ID']
  user: User
  userId: Scalars['ID']
}

export type VatType = {
  __typename?: 'VatType'
  createdAt: Scalars['DateTime']
  id: Scalars['ID']
  name: Scalars['String']
  percent: Scalars['Float']
  instanceId: Scalars['ID']
  updatedAt?: Maybe<Scalars['DateTime']>
}
// ...
export type Webhook = {
  __typename?: 'Webhook'
  concern: Scalars['String']
  createdAt?: Maybe<Scalars['DateTime']>
  event: Scalars['String']
  graphqlQuery?: Maybe<Scalars['String']>
  headers?: Maybe<Array<WebhookHeader>>
  id: Scalars['ID']
  lastInvocation?: Maybe<WebhookInvocation>
  method: HttpMethod
  name: Scalars['String']
  pastInvocations?: Maybe<Array<WebhookInvocation>>
  instance?: Maybe<Instance>
  instanceId: Scalars['ID']
  updatedAt?: Maybe<Scalars['DateTime']>
  url: Scalars['String']
}